Restructure In Progress/src - restructuring/ClubKitApi/requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.5
django-cors-middleware==1.3.1
djangorestframework==3.9.1
idna==2.8
Pillow==5.4.1
python-decouple==3.1
pytz==2018.9
requests==2.21.0
static3==0.7.0
urllib3==1.24.1
django-cripsy-forms==1.7.2
